==> notebooks/01_data_collection_api.ipynb <==
import requests
import pandas as pd


# SpaceX API endpoint for past launches
url = "https://api.spacexdata.com/v4/launches/past"

# Sending GET request
response = requests.get(url)

# Checking response status
print("Status Code:", response.status_code)

# Converting response to JSON
launches = response.json()

# Showing number of records
len(launches)


# Converting JSON to pandas DataFrame
df = pd.json_normalize(launches)

# Displaying first 5 rows
df.head()


df = df[
    [
        'flight_number',
        'name',
        'date_utc',
        'success',
        'rocket',
        'payloads',
        'launchpad'
    ]
]

df.head()


# Converting date column to datetime
df['date_utc'] = pd.to_datetime(df['date_utc'])

# Extracting launch year
df['launch_year'] = df['date_utc'].dt.year

df.head()


# Creating class label (1 = Success, 0 = Failure)
df['class'] = df['success'].apply(lambda x: 1 if x == True else 0)

df[['flight_number', 'success', 'class']].head()


# Saving dataset to CSV
df.to_csv("spacex_launch_data_api.csv", index=False)

print("CSV file saved successfully!")


from google.colab import files

files.download("spacex_launch_data_api.csv")


# Checking dataset info
df.info()
