==> notebooks/03_data_wrangling.ipynb <==
import pandas as pd


# Loading dataset from GitHub
url = "https://raw.githubusercontent.com/pakizaim7/spacex-capstone/main/notebooks/spacex_launch_data_api.csv"

df = pd.read_csv(url)

df.head()


df.info()


# Converting date_utc to datetime
df['date_utc'] = pd.to_datetime(df['date_utc'])

df.dtypes


df.isnull().sum()


df['success'] = df['success'].fillna(False)


# Creating binary class label
df['class'] = df['success'].apply(lambda x: 1 if x == True else 0)

df[['success', 'class']].head()


df['launch_year'] = df['date_utc'].dt.year

df[['date_utc', 'launch_year']].head()


# Keeping only relevant columns
df = df[
    [
        'flight_number',
        'name',
        'launch_year',
        'success',
        'class',
        'rocket',
        'payloads',
        'launchpad'
    ]
]

df.head()


df.info()
df.shape


df.to_csv("spacex_launch_data_cleaned.csv", index=False)

print("Cleaned dataset saved successfully!")


from google.colab import files

files.download("spacex_launch_data_cleaned.csv")
